Add files via upload
The program asks the user for their credentials that are already saved. If the user enters the correct name and pin he/she will be taken to a menu where he/she can choose from a number of options. If the user enters incorrect login credentials, they will be asked for their credentials additional times, for about two more times before the program terminates. 
Once the user enters the correct credentials he/she will then choose from a number of options (4 options). If the user selects the first option, by pressing 1. He or she will be prompted to enter the amount they wish to send. Note that the user has an initial amount of 600 which he or she cannot surpass in any of the defined options. After entering the amount the user will be returned back to the menu where he or she can choose what to do. Should the user choose to exit the program, he or she will have to select 4 and after that they will be taken to a prompt where they will be needed to rate the quality of their experience from 1 to 3. 1 being poor and 3 being excellent!.
After that the user will meet a prompt to press any key to terminate the program completly.