Add files via upload
Designed a student management system that is able to register students. The system needs to capture the following information:
1) StudentID.
2)StudentName
3)Surname
4)Gender
5)EmailAddress
6)DateOfBirth
7)PhoneNumber
8)Qualification
The system needs to apply a promotion for any student who registers and is above 25 years of age. If the student is female, they will get 30% discount on their total fees and if it's a male, they will get 19% discount. The fee for the Degree qualification is R85000,00 and the Diploma qualification is R45000,00. The application displays a notification if the student gets the promotion listed above.